+++ test/pay/stripe/billable_test.rb
@@ -371,6 +371,23 @@ class Pay::Stripe::BillableTest < ActiveSupport::TestCase
     assert_equal "wechat_pay", @pay_customer.default_payment_method.type
   end
 
+  test "stripe customer metdata proc" do
+    pay_customer = pay_customers(:stripe)
+    metadata = {user_id: pay_customer.owner_id}
+    User.pay_customer_metadata = ->(pay_customer) { {user_id: pay_customer.owner_id} }
+    assert_equal metadata, Pay::Stripe::Billable.new(pay_customer).customer_metadata
+    User.pay_customer_metadata = nil
+  end
+
+  test "stripe customer metdata symbol" do
+    original_value = User.pay_customer_metadata
+    User.pay_customer_metadata = :stripe_metadata
+    pay_customer = pay_customers(:stripe)
+    metadata = {user_id: pay_customer.owner_id}
+    assert_equal metadata, Pay::Stripe::Billable.new(pay_customer).customer_metadata
+    User.pay_customer_metadata = original_value
+  end
+
   private
 
   def payment_method