+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ### Unreleased
 
+* `pay_customer` now supports a `metadata` option to save on Stripe::Customers - @excid3
+* `pay_customer` now supports a `default_payment_processor` option to automatically create a Pay::Customer record - @excid3
+
 # 3.0.24
 
 * Make payment method and charges consistent for Fake procsesor - @excid3